Rename: 더미데이터 파일 위치 수정
