+++ src/pages/DetailPage/DetailPost/DetailPost.tsx
@@ -1,5 +1,5 @@
 import styled from '@emotion/styled';
-import { DUMMY_DATA } from '../components/DummyData';
+import { DUMMY_DATA } from '../DummyData';
 import { Badge } from './Badge';
 import { DetailTimeTablePage } from './DetailTimeTablePage';
 import { PostContents } from './PostContents';
