+++ src/pages/DetailPage/DummyData.ts
